fix: 작업 2. 순환참조 제거. Error 처리 파일에서 router가 필요한 throwErrorWhenResponseNotOk과 나머지 함수를 분리하였다.
